Partner Portal: Add isolated component for licenses list header
