+++ client/jetpack-cloud/sections/partner-portal/license-list/index.tsx
@@ -1,40 +1,32 @@
 /**
  * External dependencies
  */
-import React, { PropsWithChildren, ReactElement, useCallback, useContext } from 'react';
+import React, { PropsWithChildren, ReactElement, useContext } from 'react';
 import { useSelector } from 'react-redux';
 import { useTranslate } from 'i18n-calypso';
 import page from 'page';
-import classnames from 'classnames';
 import TransitionGroup from 'react-transition-group/TransitionGroup';
 import CSSTransition from 'react-transition-group/CSSTransition';
 
 /**
  * Internal dependencies
  */
+import { LICENSES_PER_PAGE } from 'calypso/state/partner-portal/licenses/constants';
+import { addQueryArgs } from 'calypso/lib/route';
 import { Card } from '@automattic/components';
-import {
-	LicenseFilter,
-	LicenseSortDirection,
-	LicenseSortField,
-} from 'calypso/jetpack-cloud/sections/partner-portal/types';
 import { License, PaginatedItems } from 'calypso/state/partner-portal/types';
-import QueryJetpackPartnerPortalLicenses from 'calypso/components/data/query-jetpack-partner-portal-licenses';
 import {
 	getPaginatedLicenses,
 	hasFetchedLicenses,
 	isFetchingLicenses,
 } from 'calypso/state/partner-portal/licenses/selectors';
-import LicenseListItem from 'calypso/jetpack-cloud/sections/partner-portal/license-list-item';
 import LicensePreview, {
 	LicensePreviewPlaceholder,
 } from 'calypso/jetpack-cloud/sections/partner-portal/license-preview';
-import { LICENSES_PER_PAGE } from 'calypso/state/partner-portal/licenses/constants';
-import Gridicon from 'calypso/components/gridicon';
 import Pagination from 'calypso/components/pagination';
-import { addQueryArgs } from 'calypso/lib/route';
-import { internalToPublicLicenseSortField } from 'calypso/jetpack-cloud/sections/partner-portal/utils';
+import QueryJetpackPartnerPortalLicenses from 'calypso/components/data/query-jetpack-partner-portal-licenses';
 import LicenseListContext from 'calypso/jetpack-cloud/sections/partner-portal/license-list-context';
+import LicenseListHeader from 'calypso/jetpack-cloud/sections/partner-portal/license-list/header';
 
 /**
  * Style dependencies
@@ -48,65 +40,13 @@ function setPage( pageNumber: number ): void {
 	page( addQueryArgs( queryParams, currentPath ) );
 }
 
-function setSortingConfig(
-	newSortField: LicenseSortField,
-	sortField: LicenseSortField,
-	sortDirection: LicenseSortDirection
-): void {
-	let direction = LicenseSortDirection.Descending;
-
-	if ( newSortField === sortField && sortDirection === direction ) {
-		direction = LicenseSortDirection.Ascending;
-	}
-
-	const queryParams = {
-		sort_field: internalToPublicLicenseSortField( newSortField ),
-		sort_direction: direction,
-		page: 1,
-	};
-	const currentPath = window.location.pathname + window.location.search;
-
-	page( addQueryArgs( queryParams, currentPath ) );
-}
-
 interface LicenseTransitionProps {
 	key?: string;
 }
 
-const LicenseTransition = ( props: React.PropsWithChildren< LicenseTransitionProps > ) => (
+const LicenseTransition = ( props: PropsWithChildren< LicenseTransitionProps > ) => (
 	<CSSTransition { ...props } classNames="license-list__license-transition" timeout={ 150 } />
 );
-
-interface SortButtonProps {
-	sortField: LicenseSortField;
-	currentSortField: LicenseSortField;
-	currentSortDirection: LicenseSortDirection;
-}
-
-function SortButton( {
-	sortField,
-	currentSortField,
-	currentSortDirection,
-	children,
-}: PropsWithChildren< SortButtonProps > ) {
-	const sort = useCallback( () => {
-		setSortingConfig( sortField, currentSortField, currentSortDirection );
-	}, [ sortField, currentSortField, currentSortDirection ] );
-
-	return (
-		<h2 className={ classnames( { 'is-selected': sortField === currentSortField } ) }>
-			<button onClick={ sort }>
-				{ children }
-				<Gridicon
-					icon="dropdown"
-					className={ classnames( 'license-list-item__sort-indicator', {
-						[ `is-sort-${ currentSortDirection }` ]: true,
-					} ) }
-				/>
-			</button>
-		</h2>
-	);
-}
 
 export default function LicenseList(): ReactElement {
 	const translate = useTranslate();
@@ -130,41 +70,7 @@ export default function LicenseList(): ReactElement {
 				page={ currentPage }
 			/>
 
-			<LicenseListItem header className="license-list__header">
-				<h2>{ translate( 'License state' ) }</h2>
-
-				<SortButton
-					sortField={ LicenseSortField.IssuedAt }
-					currentSortField={ sortField }
-					currentSortDirection={ sortDirection }
-				>
-					{ translate( 'Issued on' ) }
-				</SortButton>
-
-				{ filter !== LicenseFilter.Revoked && (
-					<SortButton
-						sortField={ LicenseSortField.AttachedAt }
-						currentSortField={ sortField }
-						currentSortDirection={ sortDirection }
-					>
-						{ translate( 'Assigned on' ) }
-					</SortButton>
-				) }
-
-				{ filter === LicenseFilter.Revoked && (
-					<SortButton
-						sortField={ LicenseSortField.RevokedAt }
-						currentSortField={ sortField }
-						currentSortDirection={ sortDirection }
-					>
-						{ translate( 'Revoked on' ) }
-					</SortButton>
-				) }
-
-				<div>{ /* Intentionally empty header. */ }</div>
-
-				<div>{ /* Intentionally empty header. */ }</div>
-			</LicenseListItem>
+			<LicenseListHeader />
 
 			<TransitionGroup className="license-list__transition-group">
 				{ showLicenses &&
